Moved LayerData & Decorator outside of backend plugin
